requirements.txt
kafka-python
requests
snowflake-connector-python
pandas
logging
pytz
